host/volume/zfs: Import file vdevs if existing.
This makes it possible for the system to easily reassemble itself after a host reboot in which zfs has forgotten about the zpools based on file vdevs.

Parts of this feature are probably a candidate for submitting patches for upstream, but it's not obvious where exactly the reusable parts of the interface would lie.

Signed-off-by: Eric Myhre <eric@flynn.io>